+++ core/src/main/java/org/vertexium/scoring/ScoringStrategy.java
@@ -1,0 +1,4 @@
+package org.vertexium.scoring;
+
+public interface ScoringStrategy {
+}